+++ tests/files/zero_sized.rs
@@ -1,8 +1,8 @@
 use lofty::ape::ApeFile;
 use lofty::flac::FlacFile;
 use lofty::iff::{AiffFile, WavFile};
-use lofty::mp3::MPEGFile;
 use lofty::mp4::Mp4File;
+use lofty::mpeg::MPEGFile;
 use lofty::AudioFile;
 
 fn read_file_with_properties<A: AudioFile>(path: &str) -> bool {